pyproject.toml
[project]
name = "memory-graph"
version = "0.0.1"
description = "A long-term memory processor and example chatbot it works with."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langgraph-checkpoint>=2.0.8",
    # Optional (for selecting different models)
    "langchain>=0.3.8",
    "python-dotenv>=1.0.1",
    "langgraph-sdk>=0.1.40",
    "langmem>=0.0.25",
    "langgraph>=0.2.6",
    "langchain-google-vertexai>=2.0.24",
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "httpx>=0.25.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.7.0",
    "unstructured[md]>=0.10.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["memory_graph", "chatbot"]
[tool.setuptools.package-dir]
"memory_graph" = "src/memory_graph"
"langgraph.templates.memory_graph" = "src/memory_graph"
"chatbot" = "src/chatbot"
"langgraph.templates.chatbot" = "src/chatbot"



[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = ["UP006", "UP007", "UP035", "D417", "E501"]
include = ["*.py", "*.pyi", "*.ipynb"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
"ntbk/*" = ["D", "UP", "T201"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_errors = true

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.2.10",
    "mypy>=1.15.0",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.2",
]

